main file
This python script runs through the screenshots in the folder specified. For each screenshot the user must enter the number of players for each team and the number of rounds each team won. The CgsOpenCv.py script is then run that uses pytesseract OCR to extract the text. The CODGODSTATS.py file then creates a JSON file with the information extracted from the image.